Add files via upload
This storytelling view offers a root-cause analysis. This view was prepared using superstore sample data. The goal of this story dashboard is to help people understand what exactly is causing losses for a certain category and in which region we are having maximum losses. The beginnings of this idea stem from the world of interesting visualizations, which are now often used in meetings as well as presentations to get across important information quickly and clearly